requirements.txt
python-dateutil
python-dotenv
django
djangorestframework
django-cors-headers
django-filter
django-extensions